Changed routes to be protocol based.
These work best with enums. They also have unique identifiers so we don't have tow different views on the view stack with the same identifier.

Use the protocol `RouteElement` for routes, these get transformed into `RouteIdentifiable` in the framework with auto generated identifier strings to be used by the router.